requirements.txt
streamlit
pandas
newspaper3k
scikit-learn